Fix a noise issue caused by random value of byteOrder
The value of byteOrder is a random value. If the random
value is 1 and it does not judge whether it gets a big-endian
key successfully at the same time, the pcm buffer will be
transferred to little-endian, which is not correct and causes
noise.

To fix this problem, we initialize the byteOrder as 0 and
judge if it gets a big-endian key

Bug: 137221067
Test: play a mp4 file with pcm audio and check if we can play
it normally
Change-Id: I709db10ce967909949c13f749869f76b1c6a80bb

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Ray <vincentrayananda@gmail.com>